+++ src/factorization.jl
@@ -764,223 +764,7 @@ function init_cacheval(alg::GenericFactorization,
     do_factorization(alg, newA, b, u)
 end
 
-# Ambiguity handling dispatch
-
 ################################## Factorizations which require solve! overloads
-
-"""
-`UMFPACKFactorization(;reuse_symbolic=true, check_pattern=true)`
-
-A fast sparse multithreaded LU-factorization which specializes on sparsity
-patterns with “more structure”.
-
-!!! note
-
-    By default, the SparseArrays.jl are implemented for efficiency by caching the
-    symbolic factorization. I.e., if `set_A` is used, it is expected that the new
-    `A` has the same sparsity pattern as the previous `A`. If this algorithm is to
-    be used in a context where that assumption does not hold, set `reuse_symbolic=false`.
-"""
-Base.@kwdef struct UMFPACKFactorization <: AbstractSparseFactorization
-    reuse_symbolic::Bool = true
-    check_pattern::Bool = true # Check factorization re-use
-end
-
-const PREALLOCATED_UMFPACK = SparseArrays.UMFPACK.UmfpackLU(SparseMatrixCSC(0, 0, [1],
-    Int[], Float64[]))
-
-function init_cacheval(alg::UMFPACKFactorization,
-        A, b, u, Pl, Pr,
-        maxiters::Int, abstol, reltol,
-        verbose::Bool, assumptions::OperatorAssumptions)
-    nothing
-end
-
-function init_cacheval(alg::UMFPACKFactorization, A::SparseMatrixCSC{Float64, Int}, b, u,
-        Pl, Pr,
-        maxiters::Int, abstol, reltol,
-        verbose::Bool, assumptions::OperatorAssumptions)
-    PREALLOCATED_UMFPACK
-end
-
-function init_cacheval(alg::UMFPACKFactorization, A::AbstractSparseArray, b, u, Pl, Pr,
-        maxiters::Int, abstol,
-        reltol,
-        verbose::Bool, assumptions::OperatorAssumptions)
-    A = convert(AbstractMatrix, A)
-    zerobased = SparseArrays.getcolptr(A)[1] == 0
-    return SparseArrays.UMFPACK.UmfpackLU(SparseMatrixCSC(size(A)..., getcolptr(A),
-        rowvals(A), nonzeros(A)))
-end
-
-function SciMLBase.solve!(cache::LinearCache, alg::UMFPACKFactorization; kwargs...)
-    A = cache.A
-    A = convert(AbstractMatrix, A)
-    if cache.isfresh
-        cacheval = @get_cacheval(cache, :UMFPACKFactorization)
-        if alg.reuse_symbolic
-            # Caches the symbolic factorization: https://github.com/JuliaLang/julia/pull/33738
-            if alg.check_pattern && pattern_changed(cacheval, A)
-                fact = lu(
-                    SparseMatrixCSC(size(A)..., getcolptr(A), rowvals(A),
-                        nonzeros(A)),
-                    check = false)
-            else
-                fact = lu!(cacheval,
-                    SparseMatrixCSC(size(A)..., getcolptr(A), rowvals(A),
-                        nonzeros(A)), check = false)
-            end
-        else
-            fact = lu(SparseMatrixCSC(size(A)..., getcolptr(A), rowvals(A), nonzeros(A)),
-                check = false)
-        end
-        cache.cacheval = fact
-        cache.isfresh = false
-    end
-
-    F = @get_cacheval(cache, :UMFPACKFactorization)
-    if F.status == SparseArrays.UMFPACK.UMFPACK_OK
-        y = ldiv!(cache.u, F, cache.b)
-        SciMLBase.build_linear_solution(alg, y, nothing, cache)
-    else
-        SciMLBase.build_linear_solution(
-            alg, cache.u, nothing, cache; retcode = ReturnCode.Infeasible)
-    end
-end
-
-"""
-`KLUFactorization(;reuse_symbolic=true, check_pattern=true)`
-
-A fast sparse LU-factorization which specializes on sparsity patterns with “less structure”.
-
-!!! note
-
-    By default, the SparseArrays.jl are implemented for efficiency by caching the
-    symbolic factorization. I.e., if `set_A` is used, it is expected that the new
-    `A` has the same sparsity pattern as the previous `A`. If this algorithm is to
-    be used in a context where that assumption does not hold, set `reuse_symbolic=false`.
-"""
-Base.@kwdef struct KLUFactorization <: AbstractSparseFactorization
-    reuse_symbolic::Bool = true
-    check_pattern::Bool = true
-end
-
-const PREALLOCATED_KLU = KLU.KLUFactorization(SparseMatrixCSC(0, 0, [1], Int[],
-    Float64[]))
-
-function init_cacheval(alg::KLUFactorization,
-        A, b, u, Pl, Pr,
-        maxiters::Int, abstol, reltol,
-        verbose::Bool, assumptions::OperatorAssumptions)
-    nothing
-end
-
-function init_cacheval(alg::KLUFactorization, A::SparseMatrixCSC{Float64, Int}, b, u, Pl,
-        Pr,
-        maxiters::Int, abstol, reltol,
-        verbose::Bool, assumptions::OperatorAssumptions)
-    PREALLOCATED_KLU
-end
-
-function init_cacheval(alg::KLUFactorization, A::AbstractSparseArray, b, u, Pl, Pr,
-        maxiters::Int, abstol,
-        reltol,
-        verbose::Bool, assumptions::OperatorAssumptions)
-    A = convert(AbstractMatrix, A)
-    return KLU.KLUFactorization(SparseMatrixCSC(size(A)..., getcolptr(A), rowvals(A),
-        nonzeros(A)))
-end
-
-# TODO: guard this against errors
-function SciMLBase.solve!(cache::LinearCache, alg::KLUFactorization; kwargs...)
-    A = cache.A
-    A = convert(AbstractMatrix, A)
-    if cache.isfresh
-        cacheval = @get_cacheval(cache, :KLUFactorization)
-        if alg.reuse_symbolic
-            if alg.check_pattern && pattern_changed(cacheval, A)
-                fact = KLU.klu(
-                    SparseMatrixCSC(size(A)..., getcolptr(A), rowvals(A),
-                        nonzeros(A)),
-                    check = false)
-            else
-                fact = KLU.klu!(cacheval, nonzeros(A), check = false)
-            end
-        else
-            # New fact each time since the sparsity pattern can change
-            # and thus it needs to reallocate
-            fact = KLU.klu(SparseMatrixCSC(size(A)..., getcolptr(A), rowvals(A),
-                nonzeros(A)))
-        end
-        cache.cacheval = fact
-        cache.isfresh = false
-    end
-    F = @get_cacheval(cache, :KLUFactorization)
-    if F.common.status == KLU.KLU_OK
-        y = ldiv!(cache.u, F, cache.b)
-        SciMLBase.build_linear_solution(alg, y, nothing, cache)
-    else
-        SciMLBase.build_linear_solution(
-            alg, cache.u, nothing, cache; retcode = ReturnCode.Infeasible)
-    end
-end
-
-## CHOLMODFactorization
-
-"""
-`CHOLMODFactorization(; shift = 0.0, perm = nothing)`
-
-A wrapper of CHOLMOD's polyalgorithm, mixing Cholesky factorization and ldlt.
-Tries cholesky for performance and retries ldlt if conditioning causes Cholesky
-to fail.
-
-Only supports sparse matrices.
-
-## Keyword Arguments
-
-  - shift: the shift argument in CHOLMOD.
-  - perm: the perm argument in CHOLMOD
-"""
-Base.@kwdef struct CHOLMODFactorization{T} <: AbstractSparseFactorization
-    shift::Float64 = 0.0
-    perm::T = nothing
-end
-
-const PREALLOCATED_CHOLMOD = cholesky(SparseMatrixCSC(0, 0, [1], Int[], Float64[]))
-
-function init_cacheval(alg::CHOLMODFactorization,
-        A, b, u, Pl, Pr,
-        maxiters::Int, abstol, reltol,
-        verbose::Bool, assumptions::OperatorAssumptions)
-    nothing
-end
-
-function init_cacheval(alg::CHOLMODFactorization,
-        A::Union{SparseMatrixCSC{T, Int}, Symmetric{T, SparseMatrixCSC{T, Int}}}, b, u,
-        Pl, Pr,
-        maxiters::Int, abstol, reltol,
-        verbose::Bool, assumptions::OperatorAssumptions) where {T <:
-                                                                Union{Float32, Float64}}
-    PREALLOCATED_CHOLMOD
-end
-
-function SciMLBase.solve!(cache::LinearCache, alg::CHOLMODFactorization; kwargs...)
-    A = cache.A
-    A = convert(AbstractMatrix, A)
-
-    if cache.isfresh
-        cacheval = @get_cacheval(cache, :CHOLMODFactorization)
-        fact = cholesky(A; check = false)
-        if !LinearAlgebra.issuccess(fact)
-            ldlt!(fact, A; check = false)
-        end
-        cache.cacheval = fact
-        cache.isfresh = false
-    end
-
-    cache.u .= @get_cacheval(cache, :CHOLMODFactorization) \ cache.b
-    SciMLBase.build_linear_solution(alg, cache.u, nothing, cache)
-end
 
 ## RFLUFactorization
